--- crates/sargon-uniffi/src/profile/v100/entities_linked_to_factor_source/accessibility/accessibility.rs
@@ -1,10 +1,0 @@
-use crate::prelude::*;
-use sargon::FactorSourceAccessibility as InternalFactorSourceAccessibility;
-
-/// An enum representing the accessibility of a factor source.
-#[derive(Clone, PartialEq, InternalConversion, uniffi::Enum)]
-pub enum FactorSourceAccessibility {
-    Device(DeviceFactorSourceAccessibility),
-
-    Ledger(LedgerHardwareWalletFactorSource),
-}
