--- crates/sargon-uniffi/src/profile/v100/entities_linked_to_factor_source/accessibility/mod.rs
@@ -1,5 +1,0 @@
-mod accessibility;
-mod device;
-
-pub use accessibility::*;
-pub use device::*;
