+++ crates/sargon-uniffi/src/profile/v100/entities_linked_to_factor_source/integrity/integrity.rs
@@ -1,0 +1,10 @@
+use crate::prelude::*;
+use sargon::FactorSourceIntegrity as InternalFactorSourceIntegrity;
+
+/// An enum representing the integrity of a factor source.
+#[derive(Clone, PartialEq, InternalConversion, uniffi::Enum)]
+pub enum FactorSourceIntegrity {
+    Device(DeviceFactorSourceIntegrity),
+
+    Ledger(LedgerHardwareWalletFactorSource),
+}
