+++ crates/sargon/src/profile/v100/entities_linked_to_factor_source/mod.rs
@@ -1,7 +1,7 @@
-mod accessibility;
 mod entities_linked_to_factor_source;
+mod integrity;
 mod profile_to_check;
 
-pub use accessibility::*;
 pub use entities_linked_to_factor_source::*;
+pub use integrity::*;
 pub use profile_to_check::*;
